Sync physics transforms at the end of the physic schedule
This gets rid of an unnecessary 1 frame delay, where we'd always see the
physics frame 1 frame behind their position in the physics world.

Also, when pausing, we'd previously never see the updates state.
